+++ package.json
@@ -4,7 +4,7 @@
   "version": "0.0.0-set-on-publish",
   "description": "",
   "scripts": {
-    "build": "del-cli dist && tsc && cpx 'packages/localnet/src/solana/**/*' dist/localnet/src/solana/ && cpx 'packages/localnet/src/sui/token/token.json' dist/localnet/src/sui/token",
+    "build": "del-cli dist && tsc && cpx 'packages/localnet/src/chains/solana/**/*' dist/localnet/src/chains/solana/ && cpx 'packages/localnet/src/chains/sui/token/token.json' dist/localnet/src/chains/sui/token",
     "lint:fix": "eslint --ext .js,.ts . --fix",
     "lint": "eslint --ext .js,.ts .",
     "localnet": "yarn build && node dist/commands/src/program.js"
@@ -28,10 +28,10 @@
       "require": "./dist/localnet/src/index.js",
       "types": "./dist/localnet/src/index.d.ts"
     },
-    "./solana/idl/gateway.json": "./dist/localnet/src/solana/idl/gateway.json",
-    "./solana/deploy/gateway.so": "./dist/localnet/src/solana/deploy/gateway.so",
-    "./solana/deploy/gateway-keypair.json": "./dist/localnet/src/solana/deploy/gateway-keypair.json",
-    "./sui/token/token.json": "./dist/localnet/src/sui/token/token.json"
+    "./solana/idl/gateway.json": "./dist/localnet/src/chains/solana/idl/gateway.json",
+    "./solana/deploy/gateway.so": "./dist/localnet/src/chains/solana/deploy/gateway.so",
+    "./solana/deploy/gateway-keypair.json": "./dist/localnet/src/chains/solana/deploy/gateway-keypair.json",
+    "./sui/token/token.json": "./dist/localnet/src/chains/sui/token/token.json"
   },
   "keywords": [],
   "author": "ZetaChain",
