+++ packages/localnet/src/chains/solana/isSolanaAvailable.ts
