+++ packages/localnet/src/chains/solana/solanaExecute.ts
@@ -6,9 +6,9 @@ import {
 import { keccak256 } from "ethereumjs-util";
 import { AbiCoder, ethers } from "ethers";
 
-import { NetworkID } from "./constants";
-import { logger } from "./logger";
-import Gateway_IDL from "./solana/idl/gateway.json";
+import { NetworkID } from "../../constants";
+import { logger } from "../../logger";
+import Gateway_IDL from "./idl/gateway.json";
 import { payer, secp256k1KeyPairTSS as tssKeyPair } from "./solanaSetup";
 
 export const solanaExecute = async ({
