+++ packages/localnet/src/chains/solana/solanaSetup.ts
@@ -11,11 +11,11 @@ import * as os from "os";
 import path from "path";
 import util from "util";
 
-import { backgroundProcessIds } from "../../commands/src/start";
-import { MNEMONIC, NetworkID } from "./constants";
+import { backgroundProcessIds } from "../../../../commands/src/start";
+import { MNEMONIC, NetworkID } from "../../constants";
 import { isSolanaAvailable } from "./isSolanaAvailable";
-import { logger } from "./logger";
-import Gateway_IDL from "./solana/idl/gateway.json";
+import { logger } from "../../logger";
+import Gateway_IDL from "./idl/gateway.json";
 import { solanaDeposit } from "./solanaDeposit";
 import { solanaDepositAndCall } from "./solanaDepositAndCall";
