+++ packages/localnet/src/index.ts
@@ -4,7 +4,7 @@ import { InitLocalnetAddress } from "../../types/zodSchemas";
 import * as ton from "./chains/ton";
 import { anvilTestMnemonic, MNEMONIC, NetworkID } from "./constants";
 import { evmSetup } from "./evmSetup";
-import { solanaSetup } from "./solanaSetup";
+import { solanaSetup } from "./chains/solana/solanaSetup";
 import { suiSetup } from "./suiSetup";
 import { createToken } from "./tokens/createToken";
 import { zetachainCall } from "./zetachainCall";
