+++ packages/tasks/src/localnet.ts
@@ -8,7 +8,7 @@ import waitOn from "wait-on";
 import { initLocalnet } from "../../localnet/src";
 import * as ton from "../../localnet/src/chains/ton";
 import { isDockerAvailable } from "../../localnet/src/isDockerAvailable";
-import { isSolanaAvailable } from "../../localnet/src/isSolanaAvailable";
+import { isSolanaAvailable } from "../../localnet/src/chains/solana/isSolanaAvailable";
 import { isSuiAvailable } from "../../localnet/src/isSuiAvailable";
 
 const LOCALNET_JSON_FILE = "./localnet.json";
